Init of Spring Web App on MySQL/MariaDB persistance
This is a default base commit without any code changes that allows the
autoconfiguring of datasource for Spring Web App.

The default datasource is with mysql drivers, which also work seamlessly
with mariaDB.

This commit also includes other initial repository files such as README.md LICENCE.md and
the very important .gitignore

Please see README for more details and workflow.